chore: split into 3 separate change request dialogs
